Communication and register all working

+++ include/message.h
@@ -3,5 +3,5 @@
 // This will be the data structure for the received message 
 typedef struct struct_message {
   String rfid;
-  float amount;
+  double amount;
 } struct_message;
